When receiving SIGTERM, enable maintenance.
That way when the node shuts down, the group FSM has done the necessary
steps already. If the node was a primary, another node has been elected to
take over already.

In most cases (systemd, docker, kubenetes) the way we know that a local node
has been asked to get down is the SIGTERM signal, after all.