Review shutdown sequence with SIGTERM.
Make it so that even if the shutdown sequence is later interrupted, the
state on file is "shutdown in maintenance". That way, at restart we can
disable maintenance properly.

This means also that we have a case where at startup the state is
wait_maintenance, and the monitor has assigned maintenance. After that FSM
transition we then want to disable maintenance. But not before.